requirements.txt
numpy
matplotlib
imageio
Pillow
